Add CharacterDisplay, SOCharacterStats Modify HealthBar
SOCharacterStats is a scriptable object class.
It will be used to set the subsystem stats of
the character, be they a kaiju or mech.

CharacterDisplay will be used to store and modify
character data as the round progresses.

It will set up the max health for the character.
The character's max health is based on the sum of
the subsystems' health multiplied by 1.5.

There are methods that update the subsytem and main
health, as well as a invoking a delegate,
HealthAffected, to update the health bar ui.

HealthBar modified to be placed on the parent object
HealthBars instead of the individual healthbar objects.
UpdateHealthBar is subscribed to delegate
HealthBar.HealthAffected. It now takes an int array
that contains the player number, the affected player's
max heath and current health values.