Modify CharacterDisplay
Dictionary value, HeadBuffs, added. This tracks
which buff is active.

2 new methods created: SetActiveHeadBuff and
GetActiveHeadBuff.

SetActiveHeadBuff accepts a string variable,
targetBuff. HeadBuffs is then iterated through.
If targetBuff matches a key in HeadBuffs, the
value for that key-value pair is set to 1.
Otherwise, it will set to 0.

GetActiveHeadBuff iterates through HeadBuffs, and
returns the dictionary key where the key-value is 1.
In the invalid scenario where nothing is found,
it will return "None" by default.