Modify MatchFlowManager and CharacterDisplay
MatchFlowManager has a commented out section that
details the expected flow of the match. The methods
still need implementing. Some of the needed
variables have been added. A new class, Move,
will be needed - this class will hold the move's
data and properties

CharacterDisplay has TempEffects, which will
track the temporary effects that are endowed by
certain moves for the turn.