Modify CharacterDisplay
Added new methods SetTempEffectActive and
GetTempEffectActive.

SetTempEffectActive accepts a string, TargetEffect,
and an int, state, that is a truthy value. It will
use TargetEffect as a key for Dictionary variable,
TempEffects, and will set the truthy value as the new
value.

GetTempEffectActive accepts a string, TargetEffect.
TempEffects is iterated through until the
corresponding key in TempEffects is found. On
finding a valid key in TempEffects, it will
return the corresponding value. Otherwise, it will
return 0.