Modify MatchFlowManager and ActionButton
MatchFlowManager has a new method SetPlayerMove.
Currently, this accepts a SOFighterMoves variable.
It currently just prints the move's name.
This method is also subscribed to
ActionButton.FightMoveSelected

ActionButton has a new delegate, FightMoveSelected.
This is invoked in OnMouseUp. It sends the
SOFighterMoves thats attached to the button that's
clicked.

ActionButton