Modify MatchFlowManager
Add ResolveMoves and ApplyMove methods.
ResolveMoves takes _priorityOutcome and
then determines which order the moves are
applied. In the even of outcomes 0 and 1
(respectively: player moves first, and
opponent moves first), ApplyMove is called.
Further development needs to be made for
the event of outcome 2 (draws - both moves
have the same priority values).

ApplyMove looks at the move's type, and then
uses the appropriate application method.

Currently all of this just prints the move's type
to console.