Modify MatchFlowManager and CharacterDisplay
Added ApplyBuff to MatchFlowManager.
This uses the MainEffectValue as the
headBuffID and then calls
CharacterDisplay.SetActiveHeadBuff.

SetActiveHeadBuff in CharacterDisplay has been
modified to accept an integer instead. A string
array is used to convert the integer into string
for use in foreach loop that finds the correct
buffKey. There was an issue in the foreach
loop where an error was occurring:

InvalidOperationException: Collection was modified;
enumeration operation may not execute.

Fix was to add a string list, keyList, that
contains the list of keys used int HeadBuffs
dictionary, and then iterate through this list instead.