Modify MatchFlowManager
Add ModifyHealth method. This will eventually
calculate the values to be applied to the targets
health data. At the moment, this only checks
whether the move is a healing or attacking type,
and then prints an appropriate debug.

It assumes all healing moves are targetted at "self",
and all attacking moves are targetted at "opponent".