Modify MatchFlowManager
Added additional checks for ModifyHealth to check
for defensive bonuses. It also checks which body
parts are being targeted, and then only calculates
the potential damage for those targeted limbs.

However, there is an issue at the end of
ModifyHealth, where the HealthData for the characters
are not being updated as expected. Will need to
investigate further.