Modify MatchFlowManager and CharacterDisplay
Commented out GetCharacterData in CharacterDisplay.
Commented out delegates in MatchFlowManager.
These are not being used at the moment.

Added ApplyDefense method into MatchFlowManager.
This sets the TempEffects["Blocking"] value to 1.
This is then applied in ModifyHealth, when
damage calculation checks are being performed.

ModifyHealth method now checks for
TempEffects["Blocking"] status when calculating
damage. If TempEffects["Blocking"] is true, then
a further check is performed to see if the potential
damage is greater than the damage threshold. If so,
the target's gaurd is broken, and a damage penalty
is added to the potential damage. If not, then then
the block reduction bonus is applied instead.