Modify CharacterDisplay
Refactored part of ModifyHealth so that
defense calculations are in a seprate method,
ApplyDefenseReductions.

Add new method, RemoveTempEffects. This is
called after ResolveMoves, and it sets all
TempEffectsto zero. There will need to be a
special check before RemoveTempEffects is
called for the AttackFail status. AttackFail
will eventually be used for as penalty
checks on missed attacks.