Modify MatchFlowManager
New method, CalculateRawDamage, added.
Originally damage to subsystems was the
product of the attack value and the maximum
main health value. This was too high per
subsystem.

Adjusted damage calculations on subsystems
so that the raw damage is calculated per
targetted subsystem first. The total raw
damage is the sum of the total subsystem
damage, and is used in damage threshold
calculations in ApplyDefenseReductions.

Similarly, ApplyDefenseReductions also applies the
penalties per element in the damage array.