Modify MatchFlowManager and ActionButton Add GameProperties
GameProperties contains 2 public static
Dictionaries. These contain the colour palette
values that are going to be used.

MatchFlowManager has a new method UpdateButtonUI.
This gets the player'sEnergyStored value, and
which subsystems still have health i.e. are not
broken. Then it invokes the new delegate,
ButtonStatusUpdated.

ActionButton has a new method, AbleToDoMove. This
is subcsribed to MatchFlowManager.UpdateButtonUI.

It will check whether the move button should be
available for the next turn. It will set the
button as inactive if the requirements are not
met e.g. not enough available energy or a
required subsystem is broken. It will set the
button as active if all requirements are met.