Modify CharacterDisplay and MatchFlowManager
Added GenerateEnergy methods to CharacterDisplay
and MatchFlowManager.

CharacterDisplay GenerateEnergy will calculate
the amount of energy to be generated for the
character when called. It will subtract the
the applied buff's energy cost before adding
the energy to the current total energy that the
play has.

Energy related data is now in an int array,
EnergyData. This contains the total energy the
player has left, the applied buff's energy cost,
and the max energy generated per turn.

Modifed MatchFlowManager.ApplyBuff so that
it updates CharacterDisplay.EnergyData with the
applied buff's energy cost.