Add GridManager and GridTile classes
These classes will be used in managing tile
occupant information.

GridManager moves the character objects to the
required target tiles, and sets the calls
methods within GridTile that sets the tile's
occupant.

There is an issue where the UI elements are
not being repositioned correctly. The
GridTile's occupant data is correctly set.