Modify GridManager and GridTile
Potential fix for UI elements not being properly
repositioned. Character object has to be made a
child of the tile first, and then the position
values have to be set. The anchor points for
the character objects have been set to
"Center, top" so that they appear above
the tile.

Commented out debug messages in UIManager,
CharacterDisplay, MatchFlowManager.