Add CategoryButton, and CategoryButton prefab Modify ActionMenuUI
CategoryButton class contains the action category
that the button represents. The action category is
how the fight moves are grouped in the UI. This is
separate to the "move" type category used in
ActionButton and MatchFlowManager.

The button will have an enum value, MoveType.

CategoryButton has a method, OnClick, that invokes
delegate, OnCategoryButtonPressed. This will
send the integer form of the MoveType enum.

ActionMenuUI has a method, ShowActionTray. This is
subscribed to OnCategoryButtonPressed in
CategoryButton. This accepts an integer form of
and will eventually open the ActionTray object and
show the moves that are related to the sent
MoveType. Currently, it just prints a debug message.