Modify ActionMenuUI asdas dasdadsasdasdasdadsasddsds
Added new methods, ShowActionTray,
ToggleActionTray, GetRelevantActions, and
DisableTraysCascade.

ShowActionTray will enable the action tray
object, and shows the moves that are related to the
ActionCategory value the button has e.g. when the
attack button is clicked, the action tray will show
the attack actions. This method is subscribed to
CategoryButton's delegate, OnCategoryButtonPressed.

ToggleActionTray will enable or disable the action
tray. It accepts a boolean integer, and an optional
ActionCategories enum variable, ActionType. If True
it will enable the action tray object and setup the
buttons. If false, it will disable the action,
subsystem target, and direction target trays.

GetRelevantActions returns a list of the
SOFightMoves values that will be set to the buttons.

DisableTraysCascade, will disable all trays from
a specific point. It's expected that the target
trays will be disabled when the action tray is
disabled.