Modify ActionMenuUI
Added method, ToggleCategoryButtonInteractable.
This accepts an integer variable, buttonType,
that corresponds to the ActionCategories enum
when forced as an integer. The method will
set the button's interactable to false.
Otherwise, it will renable the interactable.
To reenable the interactable on all buttons,
send -1.

Added a new public variable, CategoryButtons.
This holds all the Category button objects that
are iterated through when
ToggleCategoryButtonInteractable is called.