Minor refactor for color tint set up
ActionButton contained a method for setting
up the color tint properties for a button
component. This has been placed into a separate
class, ColorTintButtonSetUp, so that the
method can be used on other buttons that
require a similar set up.

ColorTintButtonSetUp has been added to the
ActionButton prefab, and the CategoryButton
prefab.