Add ActionButtonPool
ActionButtonPool is responsible for setting the
ActionButton objects within the action tray to
the relevant actions.

ActionMenuUI has a delegate, ActionTrayActived,
that is invoked to send the relevantActions
list to ActionButtonPool.

Currently, ActionButtonPool has a method,
SetUpRelevantActionButtons, that is subscribed
to ActionTrayActived. It only prints out all of
the moves in the sent list.