Modify ActionButtonPool and ActionButton
Modify ActionButtonPool method,
SetUpRelevantActionButtons. This will now loop
through all available buttons in the pool
(maximum expected is 8) and will then set up
the button. During the loop, it will:

find the next button in the list, set
ActionButton.FigherMove to relevantActions[x],
enable the button, and set the interactable
state to active. It will do this if there are
still items in relevantActions that need
setting up.

disable the action button, if there are no
more items in relevantActions.