Add Tooltip related scripts and object
Added TooltipCanvas object to scene. This will be
responsible for the tooltip when hovering over
action menus.

TooltipSystem is responsible for the methods that
show and hide the tooltip object.

TooltipManager is responsible for setting the
contenton the tooltip object.

TooltipHover is responsible for getting the
required content from the UI object that's being
mouse-overed and sending it to TooltipManager.

Currently, it is just a prototype.