Modify GameProperties and SOFightMoves
GameProperties has a new enum, TargetType.
This is used in SOFightMoves for initialising
MainTarget. It will also be used in ActionMenuUI
when determining whether the target tray is
enabled.

SOFightMoves MainTarget has been changed to use
GameProperties.TargetType instead of the custom enum.
The major differnece is that the custom enum did not
have "None" as an option.