Modify TargetButton and ActionMenuUI
TargetButton has a new delegate,
TargetButtonClicked.

New variables added to TargetButton:
Boolean variables, IsSelected, and IsDefault.
Button variable TargettingButton.

OnClick method modified so that it does not click
the button unless IsDefault is false.

TargetButtonClicked is invoked to send the
integer of the GameProperties.SubSystem enum,
and a truthy variable for whether the subsystem
is to be targetted.

ActionMenuUI has new methods for setting up the
TargetTray object: SetUpTargetTrayDefaults,
FindCounterText, SetSubSystemTarget,
SetCounterText.