Modify GameProperties and ColorTintButtonSetUp
GameProperties has a new enum, ColorCombo,
which is used in ColorTintButtonSetUp for
setting the color pallette combination.

New dictionary in ColorTintButtonSetUp,
PalleteCombo. The key is the ColorCombo
enum from GameProperties, and the values
are an array of color values that are used
to set the transitition colors on buttons.

Modified ColorTintButtonSetUp method,
SetUpButtonColours, so that it has
ColorCombo as an input, and that the
transition colors are set using
PalleteCombo. This method is now a
public method.