Add ConfirmationButton class and object Modify ActionMenuUI
ConfirmationButton Object is the final step in
which a player confirms and sends the move to
to be used in MatchFlowManager.

ConfirmationButton has new method, SetInteractable.
This method sets the interactable state of the
ConfirmationButton object. This method is subscribed
to ActionMenuUI.AllTargetsSet. It should be enabled
once all selectable targets are confirmed.

ActionMenuUI method, ToggleButtonsInteractable,
modified to invoke new delegate, AllTargetsSet.
SetUp method also modified so that ConfirmationButton
is disabled at the start of the turn.