Modify ActionMenuUI and ConfirmationButton
ConfirmationButton has a new method,
SetSelectedMove. This method is subscribed to
ActionButton.FightMoveSelected. This sets the
SelectedMove in ConfirmationButton object so
that it is ready to be sent to MatchFlowManager.

ActionMenuUI has a new method, ResetTargetedBodyParts.
This sets all values in the TargetedBodyParts to 0.
This is called within CheckTargetTray. As it stands,
the player can select all their targets for a move, and
then browse other moves without losing current selection.
However, TargetedBodyParts wasn't being reset when
another move was selected, so additional body parts were
being targeted.