Add TargetDirectionButton class and object Modified GameProperties and ActionMenuUI.
TargetDirectionButton object is in the
drection target tray. It uses the new class,
TargetDirectionButton. The relative
direction of travel is set on the button
using the new enum in GameProperties,
TargetDirection.

TargetDirectionButton has a method, OnClick.
It currently prints a debug mesage when the
button is clicked, showing which direction
was selected, the enum WhichDirection.

ActionMenuUI was modifed as there was an
null reference error when clicking an
action that does not enable a targetting tray.

Methods that require a targetting tray now
check that ActiveTargetTray is not null
before executing.