Add methods, values to XOR direction buttons
TargetDirectionButton has new methods,
variables, and delegates to help ensure that
only one direction button is selected, and sends
the correct direction value when OnClick is
called.

ToggleSelectedState is subscribed to new delegate
CheckIsSelectedRequested. This accepts a boolean,
otherButtonIsSelected.
This will set the IsSelected value to false
if the otherButtonIsSelected is true AND
PreventSelfDisable is false. PreventSelfDisable
is used to prevent the clicked button from
setting IsSelected to false on itself.
ToggleSelectedState also sets the
ColorTintButtonSetUp to a different pallete
so that the button remains "highlighted" if
the user clicks elsewhere.

OnClick has been modified to set
PreventSelfDisable to true, and to send either
WhichDirection if IsSelected is true, or 0 if
false.

ActionMenuUI method SetMoveDirection has been
modified to enable the confirmation button
if input is not 0, or disable the
confirmation button if equal to 0. Currently,
it just prints a debug message.