Modify ActionMenuUI and ConfirmationButton
ActionMenuUI has a new delegate
RelativeDirectionConfirmed. SetMoveDirection
invokes the delegate in order set
RelativeDirection in ConfirmationButton.
AllTargetsSet is also invoked to send the
evaluated statement of whether
relativeDirection is not equal to zero.

ConfirmationButton has a new method
SetRelativeDirection that's subscribed to
ActionMenuUI RelativeDirectionConfirmed.

ActionMenuUI has a new deleagte,
RelativeDirectionConfirmed. OnClick method
has been modified to invoke this delegate.
This will be used to send the relativeDirection
value to MatchFlowManager.