Modify GameProperties, ColorTintButtonSetUp and TargetDirectionButton.
Added new enum value to ColorCombo,
TargetIsNotSelected, to GameProperties.

ColorColorTintButtonSetUp PalleteCombo has a new
key-value pair, TargetIsNotSelected. This is set so
that the highlight and normal values are the same.

TargetDirectionButton method, ToggleSelectedState,
has been modified so that it sets the button's
PalleteCombo to be TargetIsNotSelected when
IsSelected is false. It is set to TargetIsSelected
if IsSelected is true.