Modify ConfirmationButton and MatchFlowManager
Replaced RelativeDirection variable with an int
array, RelativeDirectionData. This is because
MatchFlowManager's _forwardOrBackwards array
also stores the opponent's relative direction too.

RelativeDirectionData has the target index to
overwrite (assumed to be index 0 for player), and
the relative direction value.

MatchFlowManager has a new method,
SetRelativeDirection.

It will set _forwardOrBackwards depending on the
values in RelativeDirectionData.