Modify MatchFlowManager and GridManager
Uncommented line that invokes MovementCommited.

Modified GridManager method, IsValidTileID.
There was an issue where Tile id 0 was mistaken
as an invalid target. This is due to the check in
IsValidTileID not including 0. Added an = operator
to the conditional so that 0 is included as a valid
tile id.

Character can now move around on the battle grid.