Modify ActionButton
In a similar vein to TargetDirectionButton,
ActionButton now has a set methods, delegates,
and variables so that a button's interactive
state and colour pallete are set appropriately.

New method ToggleSelectedState sets the buttons
interactive state and color pallete, depending on
the new values of IsSelected and AreRequirementsMet.
This method is subscribed to ActionButton's new
delegate, CheckIsSelectedRequested.

CheckIsSelectedRequested is invoked in
SendSelectedMove.

A bool value, PreventSelfDisable, prevents the
currently selected button from resetting itself
when CheckIsSelectedRequested is invoked.