Refactor CharacterDisplay
In preparation for the changes that allow
ActionMenu to pre-emptively disable the
corresponding button of an already enabled
buff, CharacterDisplay was refactored to
use a new enum in GameProperties.

GameProperties has a new enum, BuffTypes.
This will be used for things regarding
setting and getting buffs.

MatchFlowManager modified to use the
new enum.