Modify ActionButton and MatchFlowManager
Added a check for ActionButtons that have a
Buff/Special"-type moves assigned to it.
The button is only enabled when the
corresponding buff is active.

ActionButton method, AbleToDoMove, modified.
It will check the move's type before
starting checks to see whether the button
should be selectable.

ActionButton has a new delegate,
CurrentBuffRequested. This is called during
AbleToDoMove when the move's type is "Special".

MatchFlowManager has a new method,
CheckActivePlayerBuff, which is subscribed to
ActionButton's CurrentBuffRequested. It returns
a boolean on whether the target buff is the same
as the current active buff.