Modify GameProperties and MatchFlowManager
New dictionary added to GameProperties,
ActionTypePriority. This sets the enum
ActionType against integers. This represents
the action's type priority, and is used for
tie breaking in MatchFlowManager's
ResolvePriority.

MatchFlowManager method, ResolvePriority,
now makes additional checks in the event of
equal priority outcomes. If the move's
priorities are equal, then the move's
action type is checked. The action type
priority is gotten from GameProperties'
dictionary, ActionTypePriority.

In the event that both move and action type
priority are equal, priority will be given to
the player by default.