Modify MatchFlowManager
Modified CheckCharacterDead so that it only
checks a single fighter by playerID.
CheckCharacterDead returns either true if player
main health is zero, or all 5 subsystems' health
are zero. Otherwise returns false.

Modified ResolveMoves so that CheckCharacterDead
is called after ApplyMove is called. A check is
included before ApplyMove is called for the
character moving second, where it skips that call
if a character is dead.

Modified StartMoveResolution so that there's a
check for dead characters. If so, the endgame
block of code will play out. Otherwise, other
effects are applied.