Add CPUMoveSelect and modify GameProperties
GameProperties has new variables:
int BattleIndex to track current game state,
and dictionaries KaijuMoveSetIDs and
IDNumToActionName.

KaijuMoveSetIDs is an int-int[] pair. It contains
the set of moves available for the kaiju being
found.

IDNumToActionName is an int-string pair. The
numerical ID is converted to the move's stringID.

CPUMoveSelect will determine which move the CPU
chooses.

On SetUp, GetMoveSet is called and will fetch the
set of moves that that kaiju will have.