Modify CPUMoveSelect, SOCharacterStats, and GameProperties
Added new GameProperties enum, KaijuPersonalities,
and Dictionary, PersonalitiyModifier.

KaijuPersonalities are just a set of options that
are to be set on SOCharacterStats. These are also
used as keys in PersonalitiyModifier, and determine
which modifier values are used during CPUMoveSelect.

PersonalitiyModifier is a KaijuPersonalities-int[]
dictionary. The int arrays are the modifier values
that affect the kaiju's likelihood of selcting a
specific category of move type.

SOCharacterStats has a new KaijuPersonalities
array, Behaviours. This is set from the unity
inspector. This is the allowed modifiers that
CPUMoveSelect can switch to.

CPUMoveSelect has a new method,
AdjustRatiosBasedOnFactors. At the moment, it just
returns the probabilityModifiers array.