Modify CPUMoveSelect, MatchFlowManager and GameProperties
Modified CPUMoveSelect so that the selected
fight move is stored in a global class
variable. This is so that subsystem and
direction target methods can access it more
easily without having to pass it back and
firth from MatchFlowManager.

Modified GameProperties delegate
SOFighterMoveForIntArray to
IntArrayRequested. This requires no input, and
outputs an int array.

MatchFlowManager's GetCPUMoveAndTargets method
modified. It will check the move type first. If
it is not of ActioType.Move,
CPUSubSystemTargetsRequested will be invoked. Else,
a different delegate will be invoked.