Modify CharacterDisplay and CPUMoveSelect
Changed CharacterDisplay method,
GetSystemHealthPercentage, to return a float instead.
Due to percentage rounding errors, fractionnal values
less than 1% of the maxhp are rounded up or down.
This causes an issue if hp is less than 1%, where that
systems hp is considered at 0%. Thus it would be
skipped over during TriageTargets.

TriageTargets altered to not use Array.Sort because
the wrong index values were being recorded. Added
comparisons so that it records new index when
the lowest healthPercentage is surpassed.