Modify GameProperties, CPUMoveSelect, and MatchFlowManager classes, and Mech02 and Mech03 scriptable objects
GameProperties has a new dictionary,
PersonalitiyDirection. The key is a
KaijuPersonalities value, and the value
is the TargetDirection casted as an integer.

CPUMoveSelect has a new method,
SetRelativeDirection. This returns an int
array containing the playerID (1 for CPU),
in index 0, and the relative direction of
travel in index 1.

SetRelativeDirection uses the
CurrenBehaviour value stored in
CPUMoveSelect with the PersonalitiyDirection
dictionary to return the direction value and set
it to the int array before returning the array.
This method is subscribed to MatchFlowManager's
delegate, CPUDirectionRequested.

MatchFlowManager's method, GetCPUMoveAndTargets
now invokes CPUDirectionRequested when the move
type is of type Move. SetRelativeDirection is then
called to set the CPU's relative direction in
_forwardsOrBackwards.