Modify SubSystem object and SubSystemUI
Added background to SubSystem object.

Added ToggleAndUpdateDisplay. This will toggle
the active state of the SubSystemDisplay
object. There's a check that will update the
fill values of the subsystem parts. At the
moment it just prints a debug message.

Added SetUp. This sets up the initial state
of the SubSystem object. This is subscribed
to GameManager delegate RoundHasStarted.