Add SubSystemPart Modify SubSystemUI and ToolipHover
SubSystemPart holds variables that are used
in ToolipHover. It holds information related to
the part's health, the part's owner (playerID),
and the name of the part.

SubSystemUI now sets values within SubSystemPart,
specifically health.

ToolipHover has a new method, SetStringsByButtonType.
This will set the text in the tooltip based on the
button's type. A new enum, UIButtonType, is used
to set the button's type. This enum is used
in SetStringsByButtonType.