Modify tooltip object
Added Layout element to MoveName and
Description objects within Tooltip object.
The Layout element component allows these
elements to resize, which in turn allows the
tooltip frame to resize as appropriate based
on the amount of text.

Altered color values for SubSystemUI for
still active subsystem parts

Modify ToolipHover method
SetStringsByButtonType so that it prints the
exact health values only when viewing the
mech's subsystems. For the kaiju, it only
gives a vague status analysis.